requirement.txt
Django==1.8
Werkzeug==0.9.6
argparse==1.2.1
django-bootstrap-pagination==1.5.0
django-crispy-forms==1.4.0
django-extensions==1.3.8
djangorestframework==2.3.14
ipython==2.1.0
psycopg2==2.5.3
six==1.7.2
solrpy==0.9.6
wsgiref==0.1.2
